requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-cloud-logging==3.8.0
google-cloud-translate==3.12.1
pydantic==2.5.2
python-multipart==0.0.6
aiohttp==3.9.1
